Refactor: Encapsulate external tool access and improve clarity
- Encapsulated external tool access using adapter classes for LangExtract, MindMap, and WebSearch.
- Renamed files and classes for better clarity and consistency.
- Removed redundant webSearchService and improved adapter instantiation.
